Extract drawAuras to prevent cyclic import
